pyproject.toml
[tool.poetry]
name = "musicly"
version = "0.1.0"
description = ""
authors = ["3RB16 <muhammad3rbi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
